# api/openapi/control-plane-v1.yaml
openapi: 3.1.0
...
info:
  title: AIOps Control Plane API
  version: 1.0.0
  description: Governed operational asset management API. Runtime capabilities remain closed until their independent gates pass.
servers:
  - url: /
security:
  - oidc: []
...
paths:
  /api/v1/browser-config:
    get:
      tags: [Browser]
      operationId: getBrowserConfig
      security: []
      responses:
        "200":
          $ref: "#/components/responses/BrowserConfigOK"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/environments/{environment_id}/assets:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/EnvironmentID"
    get:
      tags: [Assets]
      operationId: listAssets
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 256
        - name: service_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
        - name: kinds
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 17
            uniqueItems: true
            items:
              $ref: "#/components/schemas/AssetKind"
        - name: source_ids
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 100
            uniqueItems: true
            items:
              $ref: "#/components/schemas/UUID"
        - name: lifecycles
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 5
            uniqueItems: true
            items:
              $ref: "#/components/schemas/AssetLifecycle"
        - name: mapping_statuses
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 3
            uniqueItems: true
            items:
              $ref: "#/components/schemas/MappingStatus"
        - name: criticalities
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 4
            uniqueItems: true
            items:
              $ref: "#/components/schemas/Criticality"
        - name: data_classifications
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 4
            uniqueItems: true
            items:
              $ref: "#/components/schemas/DataClassification"
        - name: sort
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AssetSort"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          $ref: "#/components/responses/AssetPageOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
    post:
      tags: [Assets]
      operationId: createAsset
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetRequest"
      responses:
        "201":
          $ref: "#/components/responses/AssetCreated"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "413":
          $ref: "#/components/responses/Problem413"
        "415":
          $ref: "#/components/responses/Problem415"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/environments/{environment_id}/assets/{asset_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/EnvironmentID"
      - $ref: "#/components/parameters/AssetID"
    get:
      tags: [Assets]
      operationId: getAsset
      responses:
        "200":
          $ref: "#/components/responses/AssetDetailOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
    patch:
      tags: [Assets]
      operationId: patchAsset
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchAssetRequest"
      responses:
        "200":
          $ref: "#/components/responses/AssetMutationOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "413":
          $ref: "#/components/responses/Problem413"
        "415":
          $ref: "#/components/responses/Problem415"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/environments/{environment_id}/assets/{asset_id}:quarantine:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/EnvironmentID"
      - $ref: "#/components/parameters/AssetID"
      - $ref: "#/components/parameters/IdempotencyKey"
      - $ref: "#/components/parameters/IfMatch"
    post:
      tags: [Assets]
      operationId: quarantineAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionAssetRequest"
      responses:
        "200":
          $ref: "#/components/responses/AssetMutationOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "413":
          $ref: "#/components/responses/Problem413"
        "415":
          $ref: "#/components/responses/Problem415"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/environments/{environment_id}/assets/{asset_id}:retire:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/EnvironmentID"
      - $ref: "#/components/parameters/AssetID"
      - $ref: "#/components/parameters/IdempotencyKey"
      - $ref: "#/components/parameters/IfMatch"
    post:
      tags: [Assets]
      operationId: retireAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionAssetRequest"
      responses:
        "200":
          $ref: "#/components/responses/AssetMutationOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "413":
          $ref: "#/components/responses/Problem413"
        "415":
          $ref: "#/components/responses/Problem415"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/environments/{environment_id}/asset-relations:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/EnvironmentID"
    get:
      tags: [AssetRelations]
      operationId: listAssetRelations
      parameters:
        - name: asset_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
        - name: source_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
        - name: types
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 9
            uniqueItems: true
            items:
              $ref: "#/components/schemas/RelationshipType"
        - name: statuses
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 2
            uniqueItems: true
            items:
              $ref: "#/components/schemas/RelationshipStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          $ref: "#/components/responses/AssetRelationPageOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/environments/{environment_id}/service-asset-bindings:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/EnvironmentID"
    get:
      tags: [ServiceAssetBindings]
      operationId: listServiceAssetBindings
      parameters:
        - name: service_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
        - name: asset_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
        - name: roles
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 5
            uniqueItems: true
            items:
              $ref: "#/components/schemas/BindingRole"
        - name: statuses
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 2
            uniqueItems: true
            items:
              $ref: "#/components/schemas/BindingStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          $ref: "#/components/responses/ServiceAssetBindingPageOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
    post:
      tags: [ServiceAssetBindings]
      operationId: createServiceAssetBinding
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceAssetBindingRequest"
      responses:
        "201":
          $ref: "#/components/responses/ServiceAssetBindingCreated"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "413":
          $ref: "#/components/responses/Problem413"
        "415":
          $ref: "#/components/responses/Problem415"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/environments/{environment_id}/service-asset-bindings/{binding_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/EnvironmentID"
      - $ref: "#/components/parameters/BindingID"
    delete:
      tags: [ServiceAssetBindings]
      operationId: deleteServiceAssetBinding
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteServiceAssetBindingRequest"
      responses:
        "204":
          $ref: "#/components/responses/ServiceAssetBindingDeleted"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "413":
          $ref: "#/components/responses/Problem413"
        "415":
          $ref: "#/components/responses/Problem415"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/asset-sources:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
    get:
      tags: [AssetSources]
      operationId: listAssetSources
      parameters:
        - name: kinds
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 10
            uniqueItems: true
            items:
              $ref: "#/components/schemas/SourceKind"
        - name: statuses
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 4
            uniqueItems: true
            items:
              $ref: "#/components/schemas/SourceStatus"
        - name: gate_statuses
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 5
            uniqueItems: true
            items:
              $ref: "#/components/schemas/SourceGateStatus"
        - name: usage
          in: query
          required: false
          description: Must be supplied together with environment_id.
          schema:
            type: string
            enum: [manual_asset_create]
        - name: environment_id
          in: query
          required: false
          description: Must be supplied together with usage=manual_asset_create.
          schema:
            $ref: "#/components/schemas/UUID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          $ref: "#/components/responses/AssetSourcePageOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/asset-sources/{source_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/SourceID"
    get:
      tags: [AssetSources]
      operationId: getAssetSource
      responses:
        "200":
          $ref: "#/components/responses/AssetSourceDetailOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/asset-source-runs/{run_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/RunID"
    get:
      tags: [AssetSources]
      operationId: getAssetSourceRun
      responses:
        "200":
          $ref: "#/components/responses/AssetSourceRunOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/asset-conflicts:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
    get:
      tags: [AssetConflicts]
      operationId: listAssetConflicts
      parameters:
        - name: environment_id
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: asset_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
        - name: source_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
        - name: statuses
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 3
            uniqueItems: true
            items:
              $ref: "#/components/schemas/ConflictStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          $ref: "#/components/responses/AssetConflictPageOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
  /api/v1/workspaces/{workspace_id}/asset-conflicts/{conflict_id}:resolve:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/ConflictID"
      - $ref: "#/components/parameters/IdempotencyKey"
      - $ref: "#/components/parameters/IfMatch"
    post:
      tags: [AssetConflicts]
      operationId: resolveAssetConflict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveAssetConflictRequest"
      responses:
        "200":
          $ref: "#/components/responses/AssetConflictMutationOK"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "413":
          $ref: "#/components/responses/Problem413"
        "415":
          $ref: "#/components/responses/Problem415"
        "500":
          $ref: "#/components/responses/Problem500"
        "503":
          $ref: "#/components/responses/Problem503"
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example.com/realms/aiops/.well-known/openid-configuration
  parameters:
    WorkspaceID:
      name: workspace_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    EnvironmentID:
      name: environment_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    BrowserConfigOK:
      description: Public browser bootstrap metadata.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
        X-Trace-ID:
          $ref: "#/components/headers/TraceID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BrowserConfig"
...
    Problem503:
      description: Required durable or identity dependency is unavailable.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
        X-Trace-ID:
          $ref: "#/components/headers/TraceID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    UUID:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
...
    AssetKind:
      type: string
      enum:
        - SERVICE
        - LINUX_VM
        - WINDOWS_VM
        - BARE_METAL_HOST
        - KUBERNETES_CLUSTER
        - KUBERNETES_NAMESPACE
        - KUBERNETES_WORKLOAD
        - DATABASE_INSTANCE
        - DATABASE
        - METRICS_SOURCE
        - LOG_SOURCE
        - TRACE_SOURCE
        - AWX_INVENTORY
        - ARGO_APPLICATION
        - CI_PIPELINE
        - GIT_REPOSITORY
        - CLOUD_RESOURCE
    AssetLifecycle:
      type: string
      enum: [DISCOVERED, ACTIVE, STALE, QUARANTINED, RETIRED]
    MappingStatus:
      type: string
      enum: [EXACT, AMBIGUOUS, UNRESOLVED]
    Criticality:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    DataClassification:
      type: string
      enum: [PUBLIC, INTERNAL, CONFIDENTIAL, RESTRICTED]
    AssetSort:
      type: string
      enum: [display_name_asc, last_observed_at_desc]
      default: display_name_asc
...